// src/types/authorizer/v2DecisionTreeResponse.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Authorizer
 * Authorizer service
 * OpenAPI spec version: v0.20.10
 */
import type { V2DecisionTreeResponsePath } from './v2DecisionTreeResponsePath';

export interface V2DecisionTreeResponse {
  path?: V2DecisionTreeResponsePath;
  path_root?: string;
}